feat: task CD

+++ app/src/main/java/com/mobdeve/s12/mp/gamification/localdb/TaskSkillReward.kt
@@ -16,7 +16,7 @@ import java.sql.Timestamp
 // DB TabLE
 @Entity(tableName = "task_skills", primaryKeys = ["taskId", "skillId"])
 data class TaskSkillRewardEntity (
-    @ColumnInfo(name = "taskId") val taskId: Int, // Foreign key referencing the task
+    @ColumnInfo(name = "taskId") val taskId: Long, // Foreign key referencing the task
     @ColumnInfo(name = "skillId") val skillId: Int, // Foreign key referencing the skill
 
     val reward: Float
